〈開課流程〉step1~step4之功能優化及UI優化
1.step 1新增：影片連結欄位、上傳圖片功能、預覽圖片、限制字數功能(及顯示剩餘字數)
2.step 2新增：限制字數功能(及顯示剩餘字數)
3.step 2異動：rich text editor 移除 上傳影片功能
4.step 4新增：預覽課程介紹、預覽課程內容、預覽問題與討論區
5.送出申請後，若所有資訊皆已填寫，則可直接上架
