+++ public/style.css
@@ -72,9 +72,10 @@ Reset Styles
 */
 * {
   box-sizing: border-box;
-  font-family: Microsoft JhengHei; 
+  font-family: "Microsoft JhengHei", "Sans Serif","apple-system","BlinkMacSystemFont","Segoe UI","Roboto","Helvetica Neue","Arial","sans-serif","Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol";
   /* 微軟正黑體 */
 }
+
 
 article,
 aside,
@@ -442,7 +443,7 @@ h1, h2, h3, h4, h5, h6, .h1, .h2, .h3, .h4, .h5, .h6, address, p, pre, blockquot
   margin: 0; }
 
 h1, h2, h3, h4, h5, h6, .h1, .h2, .h3, .h4, .h5, .h6 {
-  font-family: "Poppins", sans-serif;
+  font-family: "Microsoft JhengHei","Poppins", sans-serif;
   line-height: 1.35; }
 
 h1,
