+++ views/createCourse/createCourseStep3.handlebars
@@ -1,21 +1,22 @@
 <div class="row">
   <div class="col-8 mx-auto">
-    <h2 class="my-3">課程名稱：{{course.name}}</h2>
+    <h3 class="my-3">
+      課程名稱：{{#if course.name}}{{course.name}}{{else}}尚未建立{{/if}}</h3>
 
     <h3 class="my-3">Step 3</h3>
     <form action="/courses/create/{{course.id}}/step3?_method=PUT" method="POST">
       <div class="form-group">
-        <label for="price">定價</label>
+        <label for="price">定價 $</label>
         <input type="number" class="form-control" name="price" id="price" placeholder="" value="{{course.price}}">
       </div>
 
       <button type="submit" class="btn btn-primary mr-5">Save</button>
     </form>
 
-    <div class="d-flex justify-content-start mt-3">
+    <div class="d-flex justify-content-start mt-3 mb-5">
       <a href="/courses/create/{{course.id}}/step2" class="btn btn-outline-info my-3 mr-5">上一步
       </a>
-      <a href="/courses/create/{{course.id}}/step4" class="btn btn-outline-info my-3">下一步
+      <a href="/courses/create/{{course.id}}/step4/intro" class="btn btn-outline-info my-3">下一步
       </a>
     </div>
   </div>
