+++ views/createCourse/createCourseStep4-intro.handlebars
@@ -1,0 +1,177 @@
+<div class="row pb-5">
+  <div class="col-8 mx-auto">
+    <h2 class="my-3">課程名稱：{{course.name}}</h2>
+
+    <h3 class="my-3">Step 4：預覽你已建立的課程</h3>
+
+    <form action="/courses/create/{{course.id}}/step4" method="POST" class="mt-3">
+      <button type="submit" class="btn btn-primary mr-2">送出申請</button>
+      <small class="text-muted">確定無誤即可送出</small>
+      <div class="d-flex justify-content-start mt-3">
+        <a href="/courses/create/{{course.id}}/step3" class="btn btn-outline-info my-3 mr-5">上一步
+        </a>
+      </div>
+
+    </form>
+  </div>
+  {{!-- 預覽課程 --}}
+  <div class="col-lg-10 mx-auto rounded border border-secondary">
+    <div class="maincontent bg--white pt--30 pb--55">
+      <div class="container">
+        <div class="row">
+
+          <div class="col-lg-12 col-12">
+
+            <h1 class="course-name">{{course.name}}</h1>
+
+            <div class="product__info__detailed">
+              <div class="pro_details_nav nav justify-content-start" role="tablist">
+                <a class="nav-item nav-link active" href="/courses/create/{{course.id}}/step4/intro">課程介紹</a>
+                <a class="nav-item nav-link" href="/courses/create/{{course.id}}/step4/lessons">課程內容</a>
+                <a class="nav-item nav-link" href="/courses/create/{{course.id}}/step4/post">問題與討論</a>
+              </div>
+            </div>
+
+            <div class="wn__single__product">
+              <div class="row">
+                <div class="col-lg-8 col-12">
+                  <div class="course-vedio">
+                    <iframe src="{{course.introVideo}}" width="100%" height="520px" frameborder="0"
+                      allowfullscreen></iframe>
+                  </div>
+                </div>
+
+
+                <div class="col-lg-4 col-12">
+                  <div class="product__info__main">
+                    <h1>{{course.name}}</h1>
+
+                    <div class="product__overview">
+                      <p><i class="fas fa-star mr-2"></i>{{course.ratingAverage}}</p>
+                      <p><i class="far fa-comment-dots mr-2"></i>{{course.ratingCount}} ratings</p>
+                      <p><i class="fas fa-users mr-1"></i>{{course.studentCount}} students enrolled</p>
+                      <p><i class="far fa-clock mr-2"></i>{{course.totalTime}} mins</p>
+                      <p class="mr-2">Created by <strong>{{course.User.username}}</strong></p>
+                    </div>
+
+                    <h2 class="course-prize">NT$ {{course.price}}</h2>
+
+                    <div class="box-tocart d-flex">
+                      <div class="addtocart__actions">
+                        {{!-- <form action="/favorite/{{course.id}}" method="POST"> --}}
+                        <button type="submit" class="tocart">ADD TO
+                          WISHLIST</button>
+                        {{!-- </form> --}}
+                      </div>
+
+                      <div class="addtocart__actions">
+                        {{!-- <form action="/order/{{course.id}}" method="POST"> --}}
+                        <button type="submit" class="tocart">BUY
+                          NOW</button>
+                        {{!-- </form> --}}
+                      </div>
+                    </div>
+
+                    <div class="product-share">
+                      <ul>
+                        <li class="categories-title">Share :</li>
+                        <li>
+                          <a href="#">
+                            <i class="icon-social-twitter icons"></i>
+                          </a>
+                        </li>
+                        <li>
+                          <a href="#">
+                            <i class="icon-social-tumblr icons"></i>
+                          </a>
+                        </li>
+                        <li>
+                          <a href="#">
+                            <i class="icon-social-facebook icons"></i>
+                          </a>
+                        </li>
+                        <li>
+                          <a href="#">
+                            <i class="icon-social-linkedin icons"></i>
+                          </a>
+                        </li>
+                      </ul>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+
+
+            <div class="product__info__detailed">
+              <div class="pro_details_nav nav justify-content-start" role="tablist">
+                <a class="nav-item nav-link active" data-toggle="tab" href="#nav-details" role="tab">Description</a>
+                <a class="nav-item nav-link" data-toggle="tab" href="#nav-review" role="tab">Reviews</a>
+              </div>
+              <div class="tab__container">
+                <!-- Start Single Tab Content -->
+                <div class="pro__tab_label tab-pane fade show active" id="nav-details" role="tabpanel">
+                  <div class="description__attribute">
+                    <p>{{course.description}}</p>
+                  </div>
+                </div>
+                <!-- End Single Tab Content -->
+
+                <!-- Start Single Tab Content -->
+                <div class="pro__tab_label tab-pane fade" id="nav-review" role="tabpanel">
+                  <div class="review__attribute">
+                    <h1>Student Reviews</h1>
+                    <h2></h2>
+                    <div class="review__ratings__type d-flex">
+                      <div class="review-ratings">
+                        <div class="rating-summary d-flex">
+                          <span>Quality</span>
+                          <ul class="rating d-flex">
+                            <li><i class="zmdi zmdi-star"></i></li>
+                            <li><i class="zmdi zmdi-star"></i></li>
+                            <li><i class="zmdi zmdi-star"></i></li>
+                            <li class="off"><i class="zmdi zmdi-star"></i></li>
+                            <li class="off"><i class="zmdi zmdi-star"></i></li>
+                          </ul>
+                        </div>
+
+                        <div class="rating-summary d-flex">
+                          <span>Price</span>
+                          <ul class="rating d-flex">
+                            <li><i class="zmdi zmdi-star"></i></li>
+                            <li><i class="zmdi zmdi-star"></i></li>
+                            <li><i class="zmdi zmdi-star"></i></li>
+                            <li class="off"><i class="zmdi zmdi-star"></i></li>
+                            <li class="off"><i class="zmdi zmdi-star"></i></li>
+                          </ul>
+                        </div>
+                        <div class="rating-summary d-flex">
+                          <span>value</span>
+                          <ul class="rating d-flex">
+                            <li><i class="zmdi zmdi-star"></i></li>
+                            <li><i class="zmdi zmdi-star"></i></li>
+                            <li><i class="zmdi zmdi-star"></i></li>
+                            <li class="off"><i class="zmdi zmdi-star"></i></li>
+                            <li class="off"><i class="zmdi zmdi-star"></i></li>
+                          </ul>
+                        </div>
+                      </div>
+                      <div class="review-content">
+                        <p>...</p>
+                        <p>Review by ...</p>
+                        <p>Posted on ../../..</p>
+                      </div>
+                    </div>
+                  </div>
+
+                </div>
+                <!-- End Single Tab Content -->
+              </div>
+            </div>
+
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
