+++ views/createCourse/createCourseStep4-post.handlebars
@@ -1,0 +1,122 @@
+<div class="row pb-5">
+  <div class="col-8 mx-auto">
+    <h2 class="my-3">課程名稱：{{course.name}}</h2>
+
+    <h3 class="my-3">Step 4：預覽你已建立的課程</h3>
+
+    <form action="/courses/create/{{course.id}}/step4" method="POST" class="mt-3">
+      <button type="submit" class="btn btn-primary mr-2">送出申請</button>
+      <small class="text-muted">確定無誤即可送出</small>
+      <div class="d-flex justify-content-start mt-3">
+        <a href="/courses/create/{{course.id}}/step3" class="btn btn-outline-info my-3 mr-5">上一步
+        </a>
+      </div>
+
+    </form>
+  </div>
+  {{!-- 預覽課程 --}}
+  <div class="col-lg-10 mx-auto rounded border border-secondary">
+    <div class="page-blog-details pt--30 pb--55 bg--white">
+      <div class="container">
+        <div class="row">
+          <div class="col-lg-12 col-12">
+
+            <h1 class="course-name">{{courseName}}</h1>
+
+            <div class="product__info__detailed">
+              <div class="pro_details_nav nav justify-content-start" role="tablist">
+                <a class="nav-item nav-link" href="/courses/create/{{course.id}}/step4/intro">課程介紹</a>
+                <a class="nav-item nav-link" href="/courses/create/{{course.id}}/step4/lessons">課程內容</a>
+                <a class="nav-item nav-link active" href="/courses/create/{{course.id}}/step4/post">問題與討論</a>
+              </div>
+            </div>
+
+
+            <div class="blog-details content">
+              <div class="comment_respond">
+                <h3 class="reply_title">Add New Question</h3>
+                <form class="comment__form" action="/courses/{{courseId}}/post" method="POST"
+                  style="display: contents;">
+                  <fieldset disabled>
+                    <div class="input__wrapper clearfix">
+                      <div class="input__box name one--third">
+                        <input type="text" placeholder="Title" name="subject">
+                      </div>
+                    </div>
+                    <div class="input__box">
+                      <textarea placeholder="Your question here" name="message"></textarea>
+                    </div>
+                    <div class="submite__btn">
+                      <a href="javascript:;" onclick="parentNode.parentNode.submit();">Post</a>
+                      <div class="clear-fix"></div>
+                    </div>
+                  </fieldset>
+
+                </form>
+              </div>
+              <div class="comments_area">
+
+                <h3 class="comment__title">Question ({{postsLength}}) </h3>
+                {{!-- 提問問題 Start --}}
+                {{#each posts}}
+                {{!-- 問題內容 Start --}}
+                <ul class="comment__list">
+                  <li class="post_question">
+                    <div class="wn__comment">
+                      <div class="thumb">
+                        <img src="https://fakeimg.pl/500x500/" alt="comment images">
+                      </div>
+                      <div class="content">
+                        <div class="comnt__author d-block d-sm-flex">
+                          <span class="post-title">{{this.subject}}</span>
+                        </div>
+                        <span>{{this.User.username}}</span>
+                        <span>{{moment this.createdAt}}</span>
+                        <p>{{this.message}}</p>
+                      </div>
+                  </li>
+                  {{!-- 問題內容 End --}}
+                  {{#each this.DiscussReplies}}
+                  {{!-- 該問題的Reply Start --}}
+                  <li class="comment_reply">
+                    <div class="wn__comment">
+                      <div class="thumb">
+                        <img src="https://fakeimg.pl/500x500/" alt="comment images">
+                      </div>
+                      <div class="content">
+                        <span>{{this.User.username}}</span>
+                        <span>{{moment this.createdAt}}</span>
+                        <p>{{this.message}}</p>
+                      </div>
+                    </div>
+                  </li>
+                  {{!-- 該問題的Reply End --}}
+                  {{/each}}
+
+                  {{!-- 留言給該問題 Start --}}
+                  <li class="comment_reply">
+                    <div class="wn__comment">
+                      <div class="thumb">
+                        <img src="https://fakeimg.pl/500x500/" alt="comment images">
+                      </div>
+                      <form action="/courses/{{../courseId}}/post/{{this.id}}/reply" method="POST"
+                        style="display: contents;">
+                        <textarea name="message" placeholder="Your reply here"></textarea>
+                        <button type="submit">Reply</button>
+                      </form>
+                    </div>
+                  </li>
+                  {{!-- 留言給該問題 End --}}
+
+                </ul>
+                {{/each}}
+                {{!-- 提問問題 End --}}
+              </div>
+
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
