--- views/createCourse/createCourseStep4.handlebars
@@ -1,21 +1,0 @@
-<div class="row">
-  <div class="col-8 mx-auto">
-    <h2 class="my-3">課程名稱：{{course.name}}</h2>
-
-    <h3 class="my-3">Step 4：預覽你已建立的課程</h3>
-
-
-    {{!-- 暫時先空白，待觀看單一課程頁面產出後，再同步 UI --}}
-    <p>暫時先空白，待觀看單一課程頁面產出後，再同步 UI</p>
-
-
-    <form action="/courses/create/{{course.id}}/step4" method="POST" class="mt-5">
-      <button type="submit" class="mt-3 btn btn-primary mr-5">送出申請</button>
-      <div class="d-flex justify-content-start mt-3">
-        <a href="/courses/create/{{course.id}}/step3" class="btn btn-outline-info my-3 mr-5">上一步
-        </a>
-      </div>
-
-    </form>
-  </div>
-</div>
